drop 64 bit because many VSTs do only 32 bit samples

+++ Audio.Foundation.Unmanaged/Audio.Foundation.Unmanaged.h
@@ -14,7 +14,7 @@ namespace Audio
 	{
 		namespace Unmanaged
 		{
-			typedef double Sample;
+			typedef float Sample;
 		}
 	}
 }
